Allow adding attributes to published configurations
Consider that we have these three 'variant/config selection modes'
1. strict attribute matching
2. no attribute matching (traditional configuration selection)
3. relaxed attribute matching with fallback

Both the consumer and the producer have an opinion if attribute
matching should be used or not. For published producers, the decision
was based on the fact if the produces was able to provide variants
(which are technically still configurations with attributes) or not.
With this change, the produce now explicitly states if it requires
attribute matching or not. This is interesting for the case, where
the consumer wants to do attribute matching, but the producer does not.
This is case (3), which so far was handled only by the fallback.

With this change:
- The producer can still provide variants, even if the attribute
  matching was triggered by the consumer. In practice, it will provide
  the configurations (each traditional configuration becomes a
  variant).
  This allows users to modify the attributes of these variants in
  component metadata rules and influence the matching by that.
- Because of this, the default fallback behavior needed some adjustment.
  We still want to keep it for backward compatibility. I.e. if we are in
  case (3), we want to select the fallback (usually 'default') if
  matching does not give a clear result. Thus we now also fall back in
  cases we have an ambiguous matching result. But, we never fall back
  anymore if we are in case (1).

Signed-off-by: Jendrik Johannes <jendrik@gradle.com>